requirements.txt
python-telegram-bot==13.15
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
Pillow==10.2.0
python-dotenv==1.0.0
